v6.7
also there is no way to add more lists, still just the 2 tabs for rensi and custom at the top. if you export a list it should be from that page not in the list itself on custom - your exporting the whole cannon not a single list from it.

also, the import should be on the main page because it shouild be a whole cannon also not a list. we can maybe leave that so people can quickly rebuild some from just lists not whole cannons but i want whole cannon export too.
and when is added it adds a new button with rensei/cutom/newOne/NextOne
and maybe a empty [New+] one to make another new on instead of custom. when you press it it just makes a new one with the title custom. then u can change the main name of the cannon list.

also im thinking that maybe we seperate the auto built list someone from the others. idk yet, maybe the main cannon and all others are on the left of page and its on the right when its toggledd on or something and the main one just is the way it is now when its toggled off taking up like 75% or w.e its at.

also lets put the title first not the author for mobile. also make it so if its mobile if the title/author pat gets cut off side scroll is activated and it dosent give a ... that is important for vertical scrll that will be the check boxes we added/ also make those toggle button not chec box. easier to click